big-bang: Lean-gate polarity INVERSION (Lean unconditional; one no-lean-link platform gate) + triage bulk test-greening
INVERSION (build-infra, approved): the seven per-crate Lean feature gates
(captp/coord lean-gate, federation lean-admission, turn lean-shadow,
intent verified-settle, sdk/app-framework lean-producer) are DELETED.
Lean verdicts/producers/differentials are now the unconditional default
on native; the single workspace feature no-lean-link (OFF by default)
exists only for targets that cannot link libdregg_lean.a (wasm32/zkvm,
plus a build.rs target hard-skip as defense-in-depth). node/Cargo.toml
drops all five feature enables — absence of verification is now visible
in the build graph. Verified: workspace native cargo check green (9
crates, exit 0) AND dregg-wasm wasm32-unknown-unknown green (exit 0).

TRIAGE GREENING (Opus lane, mirrored green on persvati, 6888 tests):
signed-wells i64 straggler sweep across test files; dead tests deleted
(captp_effects_pipeline, cross_federation_captp_turn, proof_obligation
example); coverage_state_constraints added; ratchet to 9.

+++ teasting/tests/cap_hygiene_r7_b2.rs
@@ -53,7 +53,7 @@ fn make_cell(seed: u8, balance: u64) -> Cell {
     pk[0] = seed;
     pk[31] = seed.wrapping_mul(37).wrapping_add(1);
     let token_id = [seed.wrapping_add(100); 32];
-    let mut cell = Cell::with_balance(pk, token_id, balance);
+    let mut cell = Cell::with_balance(pk, token_id, balance as i64);
     cell.permissions = open_permissions();
     cell
 }
